+++ buildSrc/CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- A new `SupportedVariants` enumeration type.
-- A new `jvm.variants` build configuration property.
+- A `SupportedVariants` enumeration type.
+- A `jvm.variants` build configuration property.
+- A `jvmName` DSL helper method that works with the `SupportedVariants` enumeration type.
 
 ## [1.1.0] - 2023-05-14