+++ buildSrc/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A `jvmName` DSL helper method that works with the `SupportedVariants` enumeration type.
 - A `jvmMain` and `jvmTest` DSL helper methods taking a `Named` target.
 
+### Fixed
+
+- Use `jvm` for the Kotlin/JVM documentation.
+
 ## [1.1.0] - 2023-05-14
 
 ### Added