+++ buildSrc/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - A `SupportedVariants` enumeration type.
 - `jvm.variants` and `konan.variants` build configuration properties.
-- A `jvmName` DSL helper method that works with the `SupportedVariants` enumeration type.
+- A `jvmPublication` DSL helper method that works with the `SupportedVariants` enumeration type.
 - `jvmMain` and `jvmTest` DSL helper methods taking a `Named` target.
 
 ### Fixed