+++ buildSrc/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Use `jvm` for the Kotlin/JVM documentation.
+- Use `native` for the Kotlin/Native documentation.
 
 ## [1.1.0] - 2023-05-14