+++ buildSrc/CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A `SupportedVariants` enumeration type.
 - `jvm.variants` and `konan.variants` build configuration properties.
 - `jvmMain` and `jvmTest` DSL helper methods taking a `Named` target.
-- Table of contents in the readme file.
+- Table of contents and documentation links in the readme file.
 
 ### Fixed